Updated the each program dropdown
