--- src/app/programs/all-programs/page.tsx
@@ -1,9 +1,0 @@
-import { AllPrograms } from '@/components/progams/allPrograms'
-import React from 'react'
-
-const page = () => {
-  return (<AllPrograms />
-  )
-}
-
-export default page
